Added Single Play Battle Mode + Match Play Lobby
There are still a lot of bugs that need to be fixed.
Some things are not read from or written to the database yet.
Some packets or not fully identified yet

+++ Database/Models/ChallengeProgress.cs
@@ -10,7 +10,7 @@
 
         public short Success { get; set; }
 
-        public byte Attempts { get; set; }
+        public short Attempts { get; set; }
 
         [ForeignKey("ChallengeId")]
         public Challenge Challenge { get; set; }
